+++ include/CachePolicy.h
@@ -14,13 +14,13 @@ private:
     uint32_t _sz;
     std::vector< std::optional<uint32_t> > _rooms;
 public:
-    explicit CachePolicy(uint32_t sz): _sz(sz), _rooms(sz) {}
-    [[nodiscard]] uint32_t Size() const { return _sz; }
-    std::optional<uint32_t> GetRoom(uint32_t index) { return _rooms[index]; }
-    void SetRoom(uint32_t index, uint32_t value) { _rooms[index] = value; }
+    explicit CachePolicy(uint64_t sz): _sz(sz), _rooms(sz) {}
+    [[nodiscard]] uint64_t Size() const { return _sz; }
+    std::optional<uint64_t> GetRoom(uint64_t index) { return _rooms[index]; }
+    void SetRoom(uint64_t index, uint64_t value) { _rooms[index] = value; }
 
-    virtual std::optional<uint32_t> Get(uint32_t n) = 0;
-    virtual std::pair<uint32_t, std::optional<uint32_t>> Load(uint32_t n) = 0;
+    virtual std::optional<uint64_t> Get(uint64_t n) = 0;
+    virtual std::pair<uint64_t, std::optional<uint64_t>> Load(uint64_t n) = 0;
 };
 
 #endif //KEYVALUE_CACHEPOLICY_H
